Introduce FakeAuthRepository for easier testing
